refactor: split out chromium-handler and node-handler
